Feat/add negative tests

+++ server/resolvers/course.ts
@@ -1,5 +1,7 @@
 //import bcrypt from 'bcryptjs';
 
+import { GraphQLError } from 'graphql';
+
 export const resolvers = {
   Query: {
     async courses(root, args, { models }) {
@@ -15,6 +17,13 @@ export const resolvers = {
     },
     async updateCourse(root, { id, name, points }, { models }) {
       const course = await models.Course.findByPk(id);
+      if (!course)
+        throw new GraphQLError('Course not found', {
+          extensions: {
+            code: 'COURSE_NOT_FOUND',
+            http: { status: 500 },
+          },
+        });
       course.set({
         name: name,
         points: points,
@@ -22,8 +31,16 @@ export const resolvers = {
       return course.save();
     },
     async deleteCourse(root, { id }, { models }) {
+      const course = await models.Course.findByPk(id);
+      if (!course)
+        throw new GraphQLError('Course not found', {
+          extensions: {
+            code: 'COURSE_NOT_FOUND',
+            http: { status: 500 },
+          },
+        });
       models.Course.destroy({ where: { id: id } });
-      return 'Course deleted successfully';
+      return course;
     },
   },
   Course: {
