+++ server/resolvers/user.ts
@@ -1,4 +1,5 @@
 import bcrypt from 'bcryptjs';
+import { GraphQLError } from 'graphql';
 
 export const resolvers = {
   Query: {
@@ -20,6 +21,13 @@ export const resolvers = {
     },
     async updateUser(root, { id, name, email }, { models }) {
       const user = await models.User.findByPk(id);
+      if (!user)
+        throw new GraphQLError('User not found', {
+          extensions: {
+            code: 'USER_NOT_FOUND',
+            http: { status: 500 },
+          },
+        });
       user.set({
         name: name,
         email: email,
@@ -27,9 +35,18 @@ export const resolvers = {
       return user.save();
     },
     async deleteUser(root, { id }, { models }) {
-      return models.User.destroy({
+      const user = await models.User.findByPk(id);
+      if (!user)
+        throw new GraphQLError('User not found', {
+          extensions: {
+            code: 'USER_NOT_FOUND',
+            http: { status: 500 },
+          },
+        });
+      await models.User.destroy({
         where: { id },
       });
+      return user;
     },
   },
   User: {
